Update patch set 1
Patch Set 1:

@dhacker29

I kanged this from Ethan's m8 commit.  You probably need the bind/unbind that he has in his commit.

With this commit building the wlan driver built-in "just worked".

I believe that the reason it wasn't working for me was that I didn't have the wcnss-service defined or started.  Looking at your 8226 it looks like you may be missing a start wcnss-service after doing the rest of the setup.

Patch-set: 1