mondrianwifi: audio: Don't use the speaker for speaker-and-headphones
Prior to this commit, the audio was only being played on the speaker
and not on the headphones.

"Fix" that with a hammer by taking over this use-case and making it
play only over the headphones instead.

Change-Id: I2663ba01769e27bd7ee032a0bc6de3c3d95d4ee6
(cherry picked from commit 0316aea3d49b7441851d71f70eaf94437f17f41f)