Update patch set 3
Patch Set 3:

This issue is caused by an old bootloader (4.3) I am currently using that bootloader and putting it back init.rc will fix it. It is caused by a weird race condition which is  which is why qualcomm left the firmware stuff in init.rc.  some users are still using that bootloader.

Patch-set: 3