Update patch set 3
Patch Set 3:

Matt, I fully understand what seems to be needed based on the commit that you linkd.  What I don't understand is why and what the consequences are.

If the "why" is because some partitions don't always get mounted and by moving it out of the fstab we just don't notice it/them fail to mount, that's pretty shitty.

If the "why" is because something needs to happen before mounting then maybe we can fix the real problem and avoid having it crop up again in the future.

I suspect that Shareef actually understands the why and could explain it to me, but I think he is out of touch for a little bit.  That is why I have been trying to understand it myself rather than wait for him to come back.

Patch-set: 3