Update patch set 3
Patch Set 3:

Shareef, shouldn't the "wait" take care of the availability of the partition?  That's what I was expecting.

Assuming that "wait" works the same as the init "wait", the only other difference I can see if that we are asking for a "check" as well in the fstab (which isn't being done when mounted via init).

I'm not really in the mood to try to force file system errors in my /efs/ partition to see what happens in both cases though...

Patch-set: 3