mondrianwifi: POC for device symbols (camera & utext)
This commit shows a couple examples of using the libc
device global symbols commit to export a basic camera
symbol and a more complex example that provides a
compatibility symbol to a 3rd party library that is
not directly available in libc.

Using this commit (and the bionic commit), I can
revert the hexediting of libqc-opt.so that was
required for utext symbols.

NOTE:

 * I can augment the utext symbols without needing
   to link directly against libicu.

 * I punted on logging because I don' know the right
   way to do so in libc.

 * Simply adding dummy symbols or string constants
   because trivial (e.g. the camera-parameters).

Change-Id: I6eda97d98c3e00cafb67ca1cf45d41926ebaa983