Update patch set 3
Patch Set 3:

Very good point, Ethan!

This commit was just meant to show the mechanics, not actually to suggest that this was a good way to add the camera symbols.  I had to replace a nice C++ constant with a string literal, which is definitely making this worse!

The mechanics are sound though and did provide a good way to test them.

Patch-set: 3
Reviewer: Christopher R. Palmer <6351@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>
Label: Verified=0